+++ ideas/package.json
@@ -4,7 +4,7 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
-    "start": "ts-node-dev --poll src/index.ts",
+    "start": "ts-node-dev --poll src/frameworks/api/index.ts",
     "test": "jest --watch",
     "test:ci": "jest"
   },
