+++ ideas/src/app/core/entities/idea.ts
