+++ ideas/src/app/core/entities/index.ts
